Add branching flag to the clone() method
We should not copy the inference path all the time, as that causes serious
performance regressions on large libraries and on large inference paths. Unfortunately
it's not obvious in which code paths we have to branch out and in which we should not.
This commit introduces a flag, `branch_path`, which is used in inference to specify the
points where a branching out is needed by creating a copy of the inference path of an
inference context.